requirements-full.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albumentations==1.3.1
ament-index-python==1.4.0
anyio==4.12.1
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
datasets==4.5.0
Deprecated==1.3.1
dill==0.4.0
filelock==3.20.3
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.7.0
google-auth==2.48.0
google-auth-oauthlib==1.2.4
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hb-interfaces==1.10.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.25.2
huggingface_hub==1.3.7
idna==3.11
ImageIO==2.37.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==2.15.0
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
linkattacher-msgs==0.0.0
Markdown==3.10.1
MarkupSafe==3.0.3
matplotlib==3.7.2
ml-dtypes==0.2.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.6.1
numpy==1.26.4
oauthlib==3.3.1
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging @ file:///home/task_176104885106445/conda-bld/packaging_1761049078006/work
pandas==3.0.0
Pillow==10.0.0
propcache==0.4.1
protobuf==4.25.8
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-telegram-bot==20.7
PyYAML==6.0.3
qudida==0.0.4
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-image==0.24.0
scikit-learn==1.3.0
scipy==1.17.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.3.0
tf-keras-vis==0.8.5
threadpoolctl==3.6.0
tifffile==2026.1.28
tqdm==4.67.2
typer-slim==0.21.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
wrapt==1.14.2
xacro==2.1.1
xxhash==3.6.0
yarl==1.22.0
